code/api/requirements.txt
absl-py==0.13.0
alpaca-trade-api==1.2.3
astunparse==1.6.3
boto3==1.17.111
botocore==1.20.111
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dill==0.2.9
Flask==2.0.1
Flask-Classful==0.14.2
flatbuffers==1.12
future==0.16.0
gapic-google-cloud-language-v1==0.15.3
gast==0.4.0
google-api-core==1.30.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-cloud-core==1.7.1
google-cloud-language==2.2.0
google-gax==0.15.16
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==1.0.2
newsapi-python==0.2.6
numpy==1.19.5
oauth2client==3.0.0
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.0
ply==3.8
proto-google-cloud-language-v1==0.15.3
proto-plus==1.19.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.4
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.7.0
sendgrid==6.7.1
six==1.15.0
SQLAlchemy==1.4.20
starkbank-ecdsa==1.1.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
tqdm==4.61.1
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.6
websocket-client==1.1.0
websockets==9.1
Werkzeug==2.0.1
wrapt==1.12.1